refactor(pipeline-job): fix audit violations per .agent/rules (V-001)
