+++ backend/src/main/java/com/init/pipelinejob/infrastructure/airflow/AirflowDomainPackGenerationTriggerAdapter.java
@@ -1,6 +1,5 @@
 package com.init.pipelinejob.infrastructure.airflow;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.init.pipelinejob.application.DomainPackGenerationTriggerCommand;
@@ -9,17 +8,13 @@ import com.init.pipelinejob.application.DomainPackGenerationTriggerResult;
 import com.init.pipelinejob.application.exception.AirflowConfigurationInvalidException;
 import com.init.pipelinejob.application.exception.AirflowTriggerFailedException;
 import com.init.shared.infrastructure.airflow.AirflowApiProperties;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.time.Clock;
-import java.time.Duration;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.http.converter.HttpMessageConversionException;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.ResourceAccessException;
@@ -28,22 +23,20 @@ import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestClientResponseException;
 
 @Component
-public class AirflowDomainPackGenerationTriggerAdapter implements DomainPackGenerationTriggerPort {
+public class AirflowDomainPackGenerationTriggerAdapter extends AirflowHttpSupport
+    implements DomainPackGenerationTriggerPort {
 
   private static final Logger log =
       LoggerFactory.getLogger(AirflowDomainPackGenerationTriggerAdapter.class);
-  private static final String LOG_FMT_DAG_RUN_ID = ", dagRunId={}";
-
-  private final AirflowApiProperties properties;
-  private final ObjectMapper objectMapper;
-  private final Clock clock;
-  private RestClient restClient;
 
   public AirflowDomainPackGenerationTriggerAdapter(
       AirflowApiProperties properties, ObjectMapper objectMapper, Clock clock) {
-    this.properties = properties;
-    this.objectMapper = objectMapper;
-    this.clock = clock;
+    super(properties, objectMapper, clock);
+  }
+
+  @Override
+  protected String triggerFailureMessage() {
+    return "Domain Pack Generation DAG 실행 요청에 실패했습니다.";
   }
 
   @Override
@@ -64,11 +57,11 @@ public class AirflowDomainPackGenerationTriggerAdapter implements DomainPackGene
     String dagId = dagId();
     String dagRunId = command.dagRunId();
     Long pipelineJobId = command.pipelineJobId();
-    RestClient restClient = restClient();
-    String token = requestToken(restClient, pipelineJobId);
+    RestClient client = restClient();
+    String token = requestToken(client, pipelineJobId);
 
     try {
-      restClient
+      client
           .post()
           .uri(apiPath("/api/v2/dags/{dagId}/dagRuns"), dagId)
           .headers(headers -> headers.setBearerAuth(token))
@@ -85,7 +78,7 @@ public class AirflowDomainPackGenerationTriggerAdapter implements DomainPackGene
           dagId,
           dagRunId,
           ex);
-      return reconcileDagRunOrThrow(restClient, token, dagId, dagRunId, pipelineJobId, ex);
+      return reconcileDagRunOrThrow(client, token, dagId, dagRunId, pipelineJobId, ex);
     } catch (RestClientResponseException ex) {
       if (ex.getStatusCode().isSameCodeAs(HttpStatus.CONFLICT)) {
         log.warn(
@@ -95,7 +88,7 @@ public class AirflowDomainPackGenerationTriggerAdapter implements DomainPackGene
             dagId,
             dagRunId,
             ex);
-        return reconcileDagRunOrThrow(restClient, token, dagId, dagRunId, pipelineJobId, ex);
+        return reconcileDagRunOrThrow(client, token, dagId, dagRunId, pipelineJobId, ex);
       }
       log.error(
           "Airflow DAG run creation failed: pipelineJobId={}, dagId={}, dagRunId={}, status={}",
@@ -113,7 +106,7 @@ public class AirflowDomainPackGenerationTriggerAdapter implements DomainPackGene
           dagId,
           dagRunId,
           ex);
-      return reconcileDagRunOrThrow(restClient, token, dagId, dagRunId, pipelineJobId, ex);
+      return reconcileDagRunOrThrow(client, token, dagId, dagRunId, pipelineJobId, ex);
     } catch (HttpMessageConversionException ex) {
       log.error(
           "Airflow DAG run creation response conversion failed: pipelineJobId={}, dagId={}"
@@ -127,14 +120,14 @@ public class AirflowDomainPackGenerationTriggerAdapter implements DomainPackGene
   }
 
   private DomainPackGenerationTriggerResult reconcileDagRunOrThrow(
-      RestClient restClient,
+      RestClient client,
       String token,
       String dagId,
       String dagRunId,
       Long pipelineJobId,
       RuntimeException triggerCause) {
     try {
-      if (dagRunExists(restClient, token, dagId, dagRunId)) {
+      if (dagRunExists(client, token, dagId, dagRunId)) {
         return triggerResult(dagId, dagRunId);
       }
     } catch (RestClientException | HttpMessageConversionException ex) {
@@ -145,47 +138,6 @@ public class AirflowDomainPackGenerationTriggerAdapter implements DomainPackGene
 
   private DomainPackGenerationTriggerResult triggerResult(String dagId, String dagRunId) {
     return new DomainPackGenerationTriggerResult(dagId, dagRunId);
-  }
-
-  private String requestToken(RestClient restClient, Long pipelineJobId) {
-    try {
-      TokenResponse response =
-          restClient
-              .post()
-              .uri(apiPath("/auth/token"))
-              .contentType(MediaType.APPLICATION_JSON)
-              .body(new TokenRequest(api().username(), api().password()))
-              .retrieve()
-              .body(TokenResponse.class);
-      if (response == null || isBlank(response.accessToken())) {
-        log.error(
-            "Airflow token response did not include access_token: pipelineJobId={}", pipelineJobId);
-        throw new AirflowTriggerFailedException(pipelineJobId);
-      }
-      return response.accessToken();
-    } catch (RestClientException | HttpMessageConversionException ex) {
-      log.error("Airflow token request failed: pipelineJobId={}", pipelineJobId, ex);
-      throw new AirflowTriggerFailedException(pipelineJobId, ex);
-    }
-  }
-
-  private boolean dagRunExists(RestClient restClient, String token, String dagId, String dagRunId) {
-    try {
-      restClient
-          .get()
-          .uri(apiPath("/api/v2/dags/{dagId}/dagRuns/{dagRunId}"), dagId, dagRunId)
-          .headers(headers -> headers.setBearerAuth(token))
-          .retrieve()
-          .toBodilessEntity();
-      return true;
-    } catch (RestClientResponseException ex) {
-      if (ex.getStatusCode().isSameCodeAs(HttpStatus.NOT_FOUND)) {
-        return false;
-      }
-      throw ex;
-    } catch (RestClientException | HttpMessageConversionException ex) {
-      throw ex;
-    }
   }
 
   private ObjectNode buildDagRunRequest(DomainPackGenerationTriggerCommand command) {
@@ -199,82 +151,4 @@ public class AirflowDomainPackGenerationTriggerAdapter implements DomainPackGene
     conf.put("pipeline_job_id", command.pipelineJobId());
     return request;
   }
-
-  private synchronized RestClient restClient() {
-    if (restClient == null) {
-      restClient = buildRestClient(api());
-    }
-    return restClient;
-  }
-
-  private RestClient buildRestClient(AirflowApiProperties.Api api) {
-    SimpleClientHttpRequestFactory requestFactory = new SimpleClientHttpRequestFactory();
-    Duration connectTimeout = api == null ? null : api.connectTimeout();
-    Duration readTimeout = api == null ? null : api.readTimeout();
-    requestFactory.setConnectTimeout(durationOrDefault(connectTimeout, Duration.ofSeconds(3)));
-    requestFactory.setReadTimeout(durationOrDefault(readTimeout, Duration.ofSeconds(10)));
-    return RestClient.builder().requestFactory(requestFactory).build();
-  }
-
-  private String apiPath(String path) {
-    return normalizeBaseUrl(api().baseUrl()) + path;
-  }
-
-  private AirflowApiProperties.Api api() {
-    AirflowApiProperties.Api api = properties.api();
-    if (api == null
-        || isBlank(api.baseUrl())
-        || isBlank(api.username())
-        || isBlank(api.password())
-        || !isOriginOnlyBaseUrl(api.baseUrl())) {
-      throw new AirflowConfigurationInvalidException();
-    }
-    return api;
-  }
-
-  private String normalizeBaseUrl(String baseUrl) {
-    URI uri = parseBaseUrl(baseUrl);
-    String origin = uri.getScheme() + "://" + uri.getRawAuthority();
-    return trimTrailingSlashes(origin);
-  }
-
-  private boolean isOriginOnlyBaseUrl(String baseUrl) {
-    URI uri = parseBaseUrl(baseUrl);
-    String scheme = uri.getScheme();
-    String path = uri.getPath();
-    return ("http".equalsIgnoreCase(scheme) || "https".equalsIgnoreCase(scheme))
-        && !isBlank(uri.getHost())
-        && uri.getUserInfo() == null
-        && (path == null || path.isBlank() || "/".equals(path))
-        && uri.getQuery() == null
-        && uri.getFragment() == null;
-  }
-
-  private URI parseBaseUrl(String baseUrl) {
-    try {
-      return new URI(trimTrailingSlashes(baseUrl));
-    } catch (URISyntaxException ex) {
-      throw new AirflowConfigurationInvalidException();
-    }
-  }
-
-  private String trimTrailingSlashes(String baseUrl) {
-    String normalized = baseUrl.trim();
-    while (normalized.endsWith("/")) {
-      normalized = normalized.substring(0, normalized.length() - 1);
-    }
-    return normalized;
-  }
-
-  private Duration durationOrDefault(Duration duration, Duration defaultValue) {
-    return duration == null ? defaultValue : duration;
-  }
-
-  private boolean isBlank(String value) {
-    return value == null || value.isBlank();
-  }
-
-  private record TokenRequest(String username, String password) {}
-
-  private record TokenResponse(@JsonProperty("access_token") String accessToken) {}
 }
