+++ backend/src/main/java/com/init/pipelinejob/infrastructure/airflow/AirflowIngestionTriggerAdapter.java
@@ -1,6 +1,5 @@
 package com.init.pipelinejob.infrastructure.airflow;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.init.pipelinejob.application.IngestionAirflowTriggerPort;
@@ -8,17 +7,13 @@ import com.init.pipelinejob.application.IngestionTriggerCommand;
 import com.init.pipelinejob.application.exception.AirflowConfigurationInvalidException;
 import com.init.pipelinejob.application.exception.AirflowTriggerFailedException;
 import com.init.shared.infrastructure.airflow.AirflowApiProperties;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.time.Clock;
-import java.time.Duration;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.http.converter.HttpMessageConversionException;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.ResourceAccessException;
@@ -27,21 +22,19 @@ import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestClientResponseException;
 
 @Component
-public class AirflowIngestionTriggerAdapter implements IngestionAirflowTriggerPort {
+public class AirflowIngestionTriggerAdapter extends AirflowHttpSupport
+    implements IngestionAirflowTriggerPort {
 
   private static final Logger log = LoggerFactory.getLogger(AirflowIngestionTriggerAdapter.class);
-  private static final String LOG_FMT_DAG_RUN_ID = ", dagRunId={}";
-
-  private final AirflowApiProperties properties;
-  private final ObjectMapper objectMapper;
-  private final Clock clock;
-  private RestClient restClient;
 
   public AirflowIngestionTriggerAdapter(
       AirflowApiProperties properties, ObjectMapper objectMapper, Clock clock) {
-    this.properties = properties;
-    this.objectMapper = objectMapper;
-    this.clock = clock;
+    super(properties, objectMapper, clock);
+  }
+
+  @Override
+  protected String triggerFailureMessage() {
+    return "Ingestion DAG 실행 요청에 실패했습니다.";
   }
 
   @Override
@@ -144,45 +137,6 @@ public class AirflowIngestionTriggerAdapter implements IngestionAirflowTriggerPo
         pipelineJobId, "Ingestion DAG 실행 요청에 실패했습니다.", triggerCause);
   }
 
-  private String requestToken(RestClient client, Long pipelineJobId) {
-    try {
-      TokenResponse response =
-          client
-              .post()
-              .uri(apiPath("/auth/token"))
-              .contentType(MediaType.APPLICATION_JSON)
-              .body(new TokenRequest(api().username(), api().password()))
-              .retrieve()
-              .body(TokenResponse.class);
-      if (response == null || isBlank(response.accessToken())) {
-        log.error(
-            "Airflow token response did not include access_token: pipelineJobId={}", pipelineJobId);
-        throw new AirflowTriggerFailedException(pipelineJobId, "Ingestion DAG 실행 요청에 실패했습니다.");
-      }
-      return response.accessToken();
-    } catch (RestClientException | HttpMessageConversionException ex) {
-      log.error("Airflow token request failed: pipelineJobId={}", pipelineJobId, ex);
-      throw new AirflowTriggerFailedException(pipelineJobId, "Ingestion DAG 실행 요청에 실패했습니다.", ex);
-    }
-  }
-
-  private boolean dagRunExists(RestClient client, String token, String dagId, String dagRunId) {
-    try {
-      client
-          .get()
-          .uri(apiPath("/api/v2/dags/{dagId}/dagRuns/{dagRunId}"), dagId, dagRunId)
-          .headers(headers -> headers.setBearerAuth(token))
-          .retrieve()
-          .toBodilessEntity();
-      return true;
-    } catch (RestClientResponseException ex) {
-      if (ex.getStatusCode().isSameCodeAs(HttpStatus.NOT_FOUND)) {
-        return false;
-      }
-      throw ex;
-    }
-  }
-
   private ObjectNode buildDagRunRequest(IngestionTriggerCommand command) {
     ObjectNode request = objectMapper.createObjectNode();
     request.put("dag_run_id", command.dagRunId());
@@ -195,82 +149,4 @@ public class AirflowIngestionTriggerAdapter implements IngestionAirflowTriggerPo
     conf.put("object_key", command.objectKey());
     return request;
   }
-
-  private synchronized RestClient restClient() {
-    if (restClient == null) {
-      restClient = buildRestClient(api());
-    }
-    return restClient;
-  }
-
-  private RestClient buildRestClient(AirflowApiProperties.Api api) {
-    SimpleClientHttpRequestFactory requestFactory = new SimpleClientHttpRequestFactory();
-    Duration connectTimeout = api == null ? null : api.connectTimeout();
-    Duration readTimeout = api == null ? null : api.readTimeout();
-    requestFactory.setConnectTimeout(durationOrDefault(connectTimeout, Duration.ofSeconds(3)));
-    requestFactory.setReadTimeout(durationOrDefault(readTimeout, Duration.ofSeconds(10)));
-    return RestClient.builder().requestFactory(requestFactory).build();
-  }
-
-  private String apiPath(String path) {
-    return normalizeBaseUrl(api().baseUrl()) + path;
-  }
-
-  private AirflowApiProperties.Api api() {
-    AirflowApiProperties.Api api = properties.api();
-    if (api == null
-        || isBlank(api.baseUrl())
-        || isBlank(api.username())
-        || isBlank(api.password())
-        || !isOriginOnlyBaseUrl(api.baseUrl())) {
-      throw new AirflowConfigurationInvalidException();
-    }
-    return api;
-  }
-
-  private String normalizeBaseUrl(String baseUrl) {
-    URI uri = parseBaseUrl(baseUrl);
-    String origin = uri.getScheme() + "://" + uri.getRawAuthority();
-    return trimTrailingSlashes(origin);
-  }
-
-  private boolean isOriginOnlyBaseUrl(String baseUrl) {
-    URI uri = parseBaseUrl(baseUrl);
-    String scheme = uri.getScheme();
-    String path = uri.getPath();
-    return ("http".equalsIgnoreCase(scheme) || "https".equalsIgnoreCase(scheme))
-        && !isBlank(uri.getHost())
-        && uri.getUserInfo() == null
-        && (path == null || path.isBlank() || "/".equals(path))
-        && uri.getQuery() == null
-        && uri.getFragment() == null;
-  }
-
-  private URI parseBaseUrl(String baseUrl) {
-    try {
-      return new URI(trimTrailingSlashes(baseUrl));
-    } catch (URISyntaxException ex) {
-      throw new AirflowConfigurationInvalidException();
-    }
-  }
-
-  private String trimTrailingSlashes(String baseUrl) {
-    String normalized = baseUrl.trim();
-    while (normalized.endsWith("/")) {
-      normalized = normalized.substring(0, normalized.length() - 1);
-    }
-    return normalized;
-  }
-
-  private Duration durationOrDefault(Duration duration, Duration defaultValue) {
-    return duration == null ? defaultValue : duration;
-  }
-
-  private boolean isBlank(String value) {
-    return value == null || value.isBlank();
-  }
-
-  private record TokenRequest(String username, String password) {}
-
-  private record TokenResponse(@JsonProperty("access_token") String accessToken) {}
 }
